requirements-docker.txt
cryptography>=3.4.7
Flask-Cors>=3.0.10
flask-marshmallow>=0.14.0
flask-restx>=0.3.0
Flask-SQLAlchemy>=2.5.1
iso3166==1.0.1
PyJWT>=2.1.0
